Update Qullamaggie exports
